--- src/App.css
@@ -1,84 +1,0 @@
-.pic{
-  margin-top:5%;
-  margin-left:5%;
-  width: 45px;
-  height: 45px;
-  border-radius: 50%;
-  background-color: lavender;
-  display: flex;
-  align-items:center;
-  justify-content:center;
-  font-size: 14px;
-  font-weight: 900;
-  color: blue;
-  font-family: 'Muli', sans-serif;
-}
-
-
-
-.date{
-  position: relative;
-  margin-top:-4%;
-  margin-left: 75%;
-  opacity: 0.5; 
-  font-size: 12px;
-  line-height: 1.85;
-  letter-spacing: 0.3px;
-  color: #4a4a4a;
-  font-family: 'Muli', sans-serif;
-}
-
-.contribution{
-    display: block;
-    margin-top:5%;
-    margin-left: 20%;
-    width: 80%;
-    height: 24px;
-    font-size: 12px;
-    line-height: 1.5;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    letter-spacing: normal;
-    color: #797979;
-    left: 10.3px;
-    font-family: 'Muli', sans-serif;
-  }
-  
-  .Loading {
-    width:100%;
-    height: 100%;
-    position: absolute;
-    margin-top: 10%;
-    display: flex;
-    justify-content: center;
-  }
-  .Loading:after {
-    content: " ";
-    display: block;
-    width: 70px;
-    height: 70px;
-    margin: 1px;
-    border-radius: 50%;
-    border: 5px solid #fff;
-    border-color: black transparent black transparent;
-    animation: lds-dual-ring 1.2s linear infinite;
-  }
-  @keyframes lds-dual-ring {
-    0% {
-      transform: rotate(0deg);
-    }
-    100% {
-      transform: rotate(360deg);
-    }
-  }
-  
-
-
-
-  @media (max-width: 550px) 
-  {
-  .contribution{
-    width:70%;
-  }
-}
