+++ src/Components/Contribution/index.css
@@ -1,0 +1,23 @@
+.contribution{
+    display: block;
+    margin-top:5%;
+    margin-left: 20%;
+    width: 80%;
+    height: 24px;
+    font-size: 12px;
+    line-height: 1.5;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    letter-spacing: normal;
+    color: #797979;
+    left: 10.3px;
+    font-family: 'Muli', sans-serif;
+  }
+  
+  @media (min-width: 550px) and (max-width: 767px){
+
+  .contribution{
+    width:80%;
+  }
+  }
