+++ src/Components/Date/index.css
@@ -1,0 +1,11 @@
+.date{
+    position: relative;
+    margin-top:-4%;
+    margin-left: 75%;
+    opacity: 0.5; 
+    font-size: 12px;
+    line-height: 1.85;
+    letter-spacing: 0.3px;
+    color: #4a4a4a;
+    font-family: 'Muli', sans-serif;
+  }
