+++ src/Components/Rectangle/index.css
@@ -1,0 +1,37 @@
+.Loading {
+    width:100%;
+    height: 100%;
+    position: absolute;
+    margin-top: 10%;
+    display: flex;
+    justify-content: center;
+  }
+  .Loading:after {
+    content: " ";
+    display: block;
+    width: 70px;
+    height: 70px;
+    margin: 1px;
+    border-radius: 50%;
+    border: 5px solid #fff;
+    border-color: black transparent black transparent;
+    animation: lds-dual-ring 1.2s linear infinite;
+  }
+  @keyframes lds-dual-ring {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
+  }
+  
+
+
+
+  @media (max-width: 550px) 
+  {
+  .contribution{
+    width:70%;
+  }
+}
